allow functions to return arbitrary length vectors
previous functions were assumed to return the same length as the input.

the new behaviour allow to return arbitrary length vectors. in case the
vector is too short, then the result is pushed to the latest date
available. in case the vector is too long, then the new dates are added.

the new behaviour allows to add functions that perform forecasts.